Move getAddableNodes to a more logical location
